+++ src/venmo/internal/ui-elements/close-icon.js
@@ -26,13 +26,13 @@ var CloseIcon = /** @class */ (function (_super) {
     }
     CloseIcon.prototype.constructElement = function () {
         var container = document.createElement("button");
-        container.id = "close-icon";
+        container.className = "close-icon";
         container.setAttribute("aria-label", "Close Venmo QR code modal");
-        container.innerHTML = "\n      <svg width=\"20\" height=\"20\" viewBox=\"0 0 20 20\" fill=\"none\" xmlns=\"http://www.w3.org/2000/svg\">\n        <path d=\"M0.949007 16.2225C0.167958 17.0036 0.167958 18.2699 0.949007 19.051C1.73006 19.832 2.99639 19.832 3.77744 19.051L9.99997 12.8284L16.2225 19.051C17.0036 19.832 18.2699 19.832 19.051 19.051C19.832 18.2699 19.832 17.0036 19.051 16.2225L12.8284 10L19.051 3.7775C19.832 2.99645 19.832 1.73012 19.051 0.949067C18.2699 0.168018 17.0036 0.16802 16.2225 0.949068L9.99997 7.17161L3.77743 0.949068C2.99639 0.16802 1.73006 0.16802 0.949007 0.949068C0.167958 1.73012 0.167959 2.99645 0.949008 3.7775L7.17155 10L0.949007 16.2225Z\" fill=\"#2F3033\"/>\n        <mask id=\"mask0\" mask-type=\"alpha\" maskUnits=\"userSpaceOnUse\" x=\"0\" y=\"0\" width=\"20\" height=\"20\">\n        <path d=\"M0.949007 16.2225C0.167958 17.0036 0.167958 18.2699 0.949007 19.051C1.73006 19.832 2.99639 19.832 3.77744 19.051L9.99997 12.8284L16.2225 19.051C17.0036 19.832 18.2699 19.832 19.051 19.051C19.832 18.2699 19.832 17.0036 19.051 16.2225L12.8284 10L19.051 3.7775C19.832 2.99645 19.832 1.73012 19.051 0.949067C18.2699 0.168018 17.0036 0.16802 16.2225 0.949068L9.99997 7.17161L3.77743 0.949068C2.99639 0.16802 1.73006 0.16802 0.949007 0.949068C0.167958 1.73012 0.167959 2.99645 0.949008 3.7775L7.17155 10L0.949007 16.2225Z\" fill=\"white\"/>\n        </mask>\n        <g mask=\"url(#mask0)\">\n        <rect width=\"20\" height=\"20\" fill=\"white\"/>\n        </g>\n      </svg>\n    ";
+        container.innerHTML = "\n      <svg width=\"20\" height=\"20\" viewBox=\"0 0 20 20\" fill=\"none\" xmlns=\"http://www.w3.org/2000/svg\">\n        <path d=\"M0.949068 16.2226C0.168019 17.0036 0.168019 18.27 0.949068 19.051C1.73012 19.8321 2.99645 19.832 3.7775 19.051L10 12.8285L16.2226 19.051C17.0036 19.8321 18.27 19.8321 19.051 19.051C19.8321 18.27 19.832 17.0036 19.051 16.2226L12.8285 10L19.051 3.77749C19.8321 2.99645 19.8321 1.73012 19.051 0.949067C18.27 0.168019 17.0036 0.16802 16.2226 0.949068L10 7.17161L3.77749 0.949068C2.99645 0.16802 1.73012 0.16802 0.949067 0.949068C0.168019 1.73012 0.16802 2.99645 0.949068 3.7775L7.17161 10L0.949068 16.2226Z\" fill=\"#6B6E76\"/>\n      </svg>\n\n    ";
         return container;
     };
     CloseIcon.prototype.getStyleConfig = function () {
-        return "\n      #close-icon {\n        cursor: pointer;\n        position: absolute;\n        top: 20px;\n        right: 20px;\n        z-index: 9999;\n        background: #2F3033;\n        box-shadow: 0px 0px 8px rgba(0, 0, 0, 0.4);\n        border-radius: 100%;\n        padding: 10px;\n      }\n\n      #close-icon svg {\n        top: 1px;\n        position: relative;\n      }\n    ";
+        return "\n      .close-icon {\n        cursor: pointer;\n        position: absolute;\n        top: 15px;\n        right: 15px;\n        z-index: 9999;\n      }\n\n      .close-icon svg {\n        top: 1px;\n        position: relative;\n      }\n    ";
     };
     return CloseIcon;
 }(base_1.default));