+++ src/venmo/shared/supports-venmo.js
@@ -5,10 +5,10 @@ var browserDetection = require('./browser-detection');
 function isBrowserSupported(options) {
   var merchantAllowsReturningToNewBrowserTab, merchantAllowsWebviews, merchantAllowsDesktopBrowsers;
   var isAndroid = browserDetection.isAndroid();
-  var isSupportedMobileDevice = isAndroid ||
-    (browserDetection.isIos() && !browserDetection.isIosChrome());
+  var isMobileDevice = isAndroid || browserDetection.isIos();
   var isAndroidChrome = isAndroid && browserDetection.isChrome();
   var isMobileDeviceThatSupportsReturnToSameTab = browserDetection.isIosSafari() || isAndroidChrome;
+  var isKnownUnsupportedMobileBrowser = browserDetection.isIosChrome() || browserDetection.isFacebookOwnedBrowserOnAndroid();
 
   options = options || {};
   // NEXT_MAJOR_VERSION allowDesktop will default to true, but can be opted out
@@ -22,27 +22,23 @@ function isBrowserSupported(options) {
   // merchant's app via a webview.
   merchantAllowsWebviews = options.hasOwnProperty('allowWebviews') ? options.allowWebviews : true;
 
+  if (isKnownUnsupportedMobileBrowser) {
+    return false;
+  }
+
   if (!merchantAllowsWebviews && (browserDetection.isAndroidWebview() || browserDetection.isIosWebview())) {
     return false;
   }
 
+  if (!isMobileDevice) {
+    return merchantAllowsDesktopBrowsers;
+  }
+
   if (!merchantAllowsReturningToNewBrowserTab) {
-    if (isMobileDeviceThatSupportsReturnToSameTab) {
-      return true;
-    }
-
-    return merchantAllowsDesktopBrowsers && !isSupportedMobileDevice;
+    return isMobileDeviceThatSupportsReturnToSameTab;
   }
 
-  if (browserDetection.isFacebookOwnedBrowserOnAndroid()) {
-    return false;
-  }
-
-  if (!merchantAllowsDesktopBrowsers) {
-    return isSupportedMobileDevice;
-  }
-
-  return true;
+  return isMobileDevice;
 }
 
 module.exports = {